refactor(render): rename server/web to server/render
The package name web collided with the existing github.com/dexidp/dex/web
(embedded frontend assets). Rename it to render, which also better describes
its job: HTML error rendering and issuer-relative URL building for the flow.
