+++ server/authflow/request_test.go
@@ -17,9 +17,9 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/dexidp/dex/server/oauth2"
+	"github.com/dexidp/dex/server/render"
 	"github.com/dexidp/dex/server/signer"
 	"github.com/dexidp/dex/server/tokens"
-	"github.com/dexidp/dex/server/web"
 	"github.com/dexidp/dex/storage"
 )
 
@@ -777,7 +777,7 @@ func TestValidateIDTokenHint(t *testing.T) {
 		signer:    sig,
 		issuerURL: *issuerURL,
 		logger:    slog.Default(),
-		UI:        &web.UI{IssuerURL: *issuerURL, Logger: slog.Default()},
+		UI:        &render.UI{IssuerURL: *issuerURL, Logger: slog.Default()},
 	}
 
 	now := time.Now()
