+++ server/logout/logout.go
@@ -14,12 +14,12 @@ import (
 	"github.com/dexidp/dex/connector"
 	"github.com/dexidp/dex/server/connectors"
 	"github.com/dexidp/dex/server/internal"
+	"github.com/dexidp/dex/server/render"
 	"github.com/dexidp/dex/server/router"
 	"github.com/dexidp/dex/server/session"
 	"github.com/dexidp/dex/server/signer"
 	"github.com/dexidp/dex/server/templates"
 	"github.com/dexidp/dex/server/tokens"
-	"github.com/dexidp/dex/server/web"
 	"github.com/dexidp/dex/storage"
 )
 
@@ -27,7 +27,7 @@ import (
 // (sessions, storage, connectors, the token issuer) and the browser rendering
 // helpers, so it carries no login-flow code.
 type Manager struct {
-	*web.UI
+	*render.UI
 
 	Storage    storage.Storage
 	Templates  *templates.Templates
